update apps to use core serializers and models
